Added IDs to events and cash machine concept in Application layer

+++ EventSourcedBank/EventSourcedBank.Data.Read/BankAccountData.cs
@@ -4,13 +4,13 @@ namespace EventSourcedBank.Data.Read
 {
     public sealed class BankAccountData
     {
-        public BankAccountData(Guid id, decimal balance)
+        public BankAccountData(Guid id, int balance)
         {
             Id = id;
             Balance = balance;
         }
 
         public Guid Id { get; }
-        public decimal Balance { get; }
+        public int Balance { get; }
     }
 }
